🎨 캐러샐 UI 스크롤 애니매이션 기능 모듈화
